+++ src/controllers/users.js
@@ -2,6 +2,7 @@ import UserModel from "../models/users";
 import jwt from "jsonwebtoken";
 import bcrypt from "bcrypt";
 import { validateBody, validateEmail } from "../utils/validators";
+import { userResponse, signToken } from "../utils/userControllerUtils"
 
 export async function registerUser(req, res) {
   try {
@@ -70,19 +71,3 @@ export async function getUser(req, res) {
   }
 }
 
-export function signToken(payload) {
-  const { token, createdAt, updatedAt, ...newPayload } = payload;
-  return jwt.sign(newPayload, process.env.PRIVATE_KEY, { algorithm: "RS256" });
-}
-
-export function userResponse(payload) {
-  return {
-    user: {
-      email: payload.email,
-      token: payload.token,
-      username: payload.username,
-      bio: payload.bio,
-      image: payload.image,
-    }
-  };
-}
